[TASK] Simplified custom torrent parser
The custom torrent parser that was in use had extra methods that were
not needed and the code was more complex than necessary. Removed the
extra methods and the inner class, and created simple Bdecode class
that handles parsing a proper torrent file.

Added test cases to make sure custom parser is completely tested
and follows the entire torrent spec.