[TASK] Updated cli and config file options doc
Updated the documentation of the cli options and the configuration file
options in the docs to match the actuals within the code.

pbr auto-generates the documentation for the code base but RTD does not
leverage the same build command. So need to include the code documentation
in source control as well.